🗃️ (DB entities): change ID types to `Long` (from `Int`)

+++ app/src/main/java/com/suvanl/fixmylinks/data/local/db/entity/AllUrlParamsRule.kt
@@ -18,8 +18,8 @@ import androidx.room.PrimaryKey
 )
 data class AllUrlParamsRule(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
 
     @ColumnInfo(name = "base_rule_id", index = true)
-    val baseRuleId: Int,
+    val baseRuleId: Long,
 )
